Clean up bab-board code

+++ hardware/bab-board/include/bab-board/battery_board.hpp
@@ -11,7 +11,7 @@
 #include "can_util/can_id_util.hpp"
 
 
-static constexpr auto VOLTAGE_MULTIPLIER = 100;
+static constexpr auto VOLTAGE_MULTIPLIER = 100.0;
 static constexpr auto CURRENT_MULTIPLIER = 100.0;
 static constexpr auto POWER_MULTIPLIER = 10.0;
 
@@ -76,23 +76,23 @@ public:
 
     // ----------------------- Battery accessors -----------------------
     // idx in [0, NUM_BATTERIES). Default 0 keeps legacy callers working.
-    float getBatteryVoltageLevel(size_t idx = 0) const;
-    float getBatteryCurrentLevel(size_t idx = 0) const;
-    float getBatteryTemp(size_t idx = 0) const;
+    double getBatteryVoltageLevel(size_t idx = 0) const;
+    double getBatteryCurrentLevel(size_t idx = 0) const;
+    double getBatteryTemp(size_t idx = 0) const;
     bool batteryFresh(size_t idx = 0, std::chrono::milliseconds max_age = std::chrono::seconds(3)) const;
     bool batteryEverReceived(size_t idx = 0) const;
 
     // ------------------------ Rail accessors -------------------------
-    float getRailVoltageLevel(size_t idx = 0) const;
-    float getRailCurrent(size_t idx = 0) const;
-    float getRailPower(size_t idx = 0) const;
-    float getRailTemp(size_t idx = 0) const; // firmware does not send; returns 0
+    double getRailVoltageLevel(size_t idx = 0) const;
+    double getRailCurrent(size_t idx = 0) const;
+    double getRailPower(size_t idx = 0) const;
+    double getRailTemp(size_t idx = 0) const; // firmware does not send; returns 0
     bool getRailSwitchOn(size_t idx = 0) const;
     bool railFresh(size_t idx = 0, std::chrono::milliseconds max_age = std::chrono::seconds(3)) const;
     bool railEverReceived(size_t idx = 0) const;
 
     // ------------------------ TCU / relay ----------------------------
-    float getTCUTemp() const;
+    double getTCUTemp() const;
     std::string getTCUStatus() const;
     std::string getRelayStatus(size_t idx = 0) const;
     bool getRelayClosed(size_t idx) const;
